requirements.txt
pandas
requests
beautifulsoup4
openpyxl
atproto
load_dotenv